add saving values in sessionStorage; extended validation for phone input
