+++ package.json
@@ -19,6 +19,7 @@
 		"react": "^18",
 		"react-dom": "^18",
 		"react-hook-form": "^7.51.4",
+		"react-hook-form-persist": "^3.0.0",
 		"react-scroll": "^1.9.0",
 		"sanity": "^3.40.0",
 		"swiper": "^11.1.1"
