+++ src/components/common/Form/FormBlock/FormBlock.types.ts
@@ -1,10 +1,3 @@
 export interface IFormBlockProps {
 	className?: string;
 }
-
-export interface IFormState {
-	name: string;
-	phone: string;
-	message: string;
-	approval: boolean;
-}
