--- src/components/common/Form/FormBlock/index.ts
@@ -1,4 +1,0 @@
-export type { IFormState } from './FormBlock.types';
-
-export { FormBlock } from './FormBlock';
-
