+++ src/data/formData.json
@@ -11,7 +11,10 @@
 		"phone": {
 			"label": "Ваш Номер телефону",
 			"placeholder": "+380967777323",
-			"errorMessage": "Недійсний номер телефону"
+			"errorMessage": "Недійсний номер телефону",
+			"minLength": "Закороткий номер. Введіть \"+380 \" потім 9 цифр. Приклад: \"+380967777323\"",
+			"maxLength": "Задовгий номер. Введіть \"+380 \" потім 9 цифр. Приклад: \"+380967777323\"",
+			"rules": "Введіть \"+380 \" потім 9 цифр. Приклад: \"+380967777323\""
 		},
 		"textarea": {
 			"label": "Ваш Коментар",
