+++ src/types/FormState.types.ts
@@ -1,0 +1,6 @@
+export interface IFormState {
+	name: string;
+	phone: string;
+	message: string;
+	approval: boolean;
+}
